deactivate notify triggers during migration and reactivate on reset
